requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.8.0
django-crontab==0.7.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.13.2
